# docs/swagger.yaml
definitions:
  errorHandler.HttpErr:
    properties:
      message:
        example: Error message
        type: string
      status_code:
        example: 400
        type: integer
    type: object
...
  input.OrganizationAddUser:
    properties:
      job:
        example: Главный по главным
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  input.OrganizationRegister:
    properties:
      address:
        type: string
      category_ids:
        items:
          type: integer
        type: array
      inn:
        type: string
      name:
        example: ЖКХ
        type: string
      phone:
        type: string
    type: object
...
  response.Organization:
    properties:
      address:
        type: string
      id:
        example: 1
        type: integer
      inn:
        type: string
      name:
        example: ЖКХ
        type: string
      phone:
        type: string
    type: object
...
  response.User:
    properties:
      email:
        example: test@gmail.com
        type: string
      first_name:
        example: Иван
        type: string
      id:
        example: 1
        type: integer
      job:
        type: string
      last_name:
        example: Иванов
        type: string
      organization:
        $ref: '#/definitions/response.Organization'
      organization_id:
        example: 1
        type: integer
      patronymic:
        example: Иванович
        type: string
    type: object
info:
  contact: {}
  description: Спецификация приложения команды подCRUDули
  title: Urbathon-2023
paths:
...
  /organization:
    post:
      consumes:
      - application/json
      description: Регистрация организации
      parameters:
      - description: OrganizationRegister
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/input.OrganizationRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Регистрация организации
      tags:
      - organization
  /organization/{id}/add_user:
    post:
      description: Добавление юзера в организацию
      parameters:
      - description: OrganizationAddUser
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/input.OrganizationAddUser'
      - description: organization id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Добавление юзера в организацию
      tags:
      - organization
swagger: "2.0"